pyproject.toml
[project]
name = "praticalcaseapi"
version = "0.1.0"
description = ""
authors = [
    {name = "zMarcoos",email = "marcosvdcdef@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "transformers (>=4.56.0,<5.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "pdfminer-six (>=20250506,<20250507)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
